fixed useUser;

+++ src/components/UtilityComponents/unfavorite-park.js
@@ -1,14 +1,16 @@
 import { usersCollection } from "../../firebase/firebase";
 
-const unfavorite = async (parkCode, user, name) => {
-  try {
-    await usersCollection
-      .doc(user[0].uid)
-      .collection("favoriteParks")
-      .doc(parkCode)
-      .delete();
-  } catch (err) {
-    console.error(err);
+const unfavorite = async (parkCode, user) => {
+  if (user) {
+    try {
+      await usersCollection
+        .doc(user.uid)
+        .collection("favoriteParks")
+        .doc(parkCode)
+        .delete();
+    } catch (err) {
+      console.error(err);
+    }
   }
 };
